chnages done in price-table-three
